% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{all_dims}
\alias{all_dims}
\title{All dims}
\usage{
all_dims()
}
\description{
This function returns an object that can be used when subsetting tensors with
`[`. If you are familiar with Python, this is equivalent to the Python Ellipsis
`...`, (not to be confused with `...` in `R`).
}
\examples{
\dontrun{
# in python, if x is a numpy array or torch tensor
x[..., i]
# the ellipsis means "expand to match number of dimension of x".
# to translate the above python expression to R, write:
x[all_dims(), i]
}
\dontrun{
# Run this
d <- torch$tensor(list(list(0, 0),
                       list(0, 0),
                       list(0, 1),
                       list(1, 1)), dtype=torch$uint8)
d[all_dims(), 1]

f <- torch$arange(9L)$reshape(c(3L, 3L))
f
f[all_dims()]
f[all_dims(), 1L]
}
}
