% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{as_boolean}
\alias{as_boolean}
\title{Convert tensor to boolean type}
\usage{
as_boolean(x)
}
\arguments{
\item{x}{a torch tensor}
}
\description{
Convert a tensor to a boolean equivalent tensor
}
\examples{
\donttest{
uo <- torch$ones(3L)
as_boolean(uo)              # tensor([True, True, True], dtype=torch.bool)
}
}
