% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{conda_install}
\alias{conda_install}
\title{Conda install}
\usage{
conda_install(envname = NULL, packages, forge = TRUE, pip = FALSE,
  pip_ignore_installed = TRUE, conda = "auto", python_version = NULL,
  channel = NULL, ...)
}
\arguments{
\item{forge}{Include the [Conda Forge](https://conda-forge.org/) repository.}

\item{pip_ignore_installed}{Ignore installed versions when using pip. This is `TRUE` by default
so that specific package versions can be installed even if they are downgrades. The `FALSE`
option is useful for situations where you don't want a pip install to attempt an overwrite
of a conda binary package (e.g. SciPy on Windows which is very difficult to install via
pip due to compilation requirements).}
}
\description{
Conda install
}
