% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset_mnist_digits}
\alias{dataset_mnist_digits}
\title{MNIST database of handwritten digits}
\usage{
dataset_mnist_digits(ntrain = 60000L, ntest = 10000L, onehot = TRUE)
}
\arguments{
\item{ntrain}{number of training samples}

\item{ntest}{number of test samples}

\item{onehot}{boolean}
}
\description{
Dataset of 60,000 28x28 grayscale images of the 10 digits, along with a test
set of 10,000 images.
}
\concept{datasets}
