% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dim.torch.Tensor}
\alias{dim.torch.Tensor}
\title{Dimensions of a tensor}
\usage{
\method{dim}{torch.Tensor}(x)
}
\arguments{
\item{x}{tensor}
}
\value{
a vector of integers with the dimensions of the tensor
}
\description{
Get the dimensions of a tensor displaying it as a vector.
}
\examples{
\dontrun{
uo = torch$ones(3L, 5L)  # it is a 3x5 tensor
dim(uo)
}

}
