% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{\%.*\%}
\alias{\%.*\%}
\title{#' @export
"round.torch.Tensor" <- function(input) {
# round: Returns a new tensor with each of the elements of input rounded to the closest integer.
torch$round(input)
}
Dot product of two tensors}
\usage{
a \%.*\% b
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
a scalar
}
\description{
This generic is similar to \code{torch$dot(a, b)}
}
\examples{
\dontrun{
p <- torch$Tensor(list(2, 3))
q <- torch$Tensor(list(2, 1))
p \%.*\% q
}

}
