% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{\%\%.torch.Tensor}
\alias{\%\%.torch.Tensor}
\title{Remainder}
\usage{
\method{\%\%}{torch.Tensor}(a, b)
}
\arguments{
\item{a}{a tensor}

\item{b}{a scalar or a tensor}
}
\value{
the reminder of the division between tensor by a scalar or tensor
}
\description{
Computes the element-wise remainder of division.
}
\examples{
\dontrun{
x <- torch$Tensor(list(-3., -2, -1, 1, 2, 3))
y <- torch$Tensor(list(1., 2, 3, 4, 5))
torch$remainder(x, 2)
torch$remainder(y, 1.5)

x \%\% 2
y \%\% 1.5
}
}
