% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{\%**\%}
\alias{\%**\%}
\title{Matrix/Tensor multiplication of two tensors}
\usage{
a \%**\% b
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\description{
This generic is similar to \code{torch$matmul(a, b)}
}
\examples{
\donttest{
p <- torch$randn(3L)
q <- torch$randn(3L)
p \%**\% q
}
}
