% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{>.torch.Tensor}
\alias{>.torch.Tensor}
\title{A tensor greater than another tensor}
\usage{
\method{>}{torch.Tensor}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\value{
A tensor of booleans representing the logical result of the comparison.
False to represent 0, and True to represent 1 in a tensor of data type \code{torch$uint8}.
}
\description{
This generic is similar to \code{torch$gt(a, b)}
}
\examples{
\donttest{
A <- torch$ones(5L, 5L)
C <- torch$as_tensor(np$random$randint(2L, size=c(5L, 5L)), dtype=torch$float32)
A > C
C > A
}
}
