% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_pytorch}
\alias{install_pytorch}
\title{Install PyTorch and its dependencies}
\usage{
install_pytorch(
  method = c("conda", "virtualenv", "auto"),
  conda = "auto",
  version = "default",
  envname = "r-torch",
  extra_packages = NULL,
  restart_session = TRUE,
  conda_python_version = "3.6",
  pip = FALSE,
  channel = "stable",
  cuda_version = NULL,
  dry_run = FALSE,
  ...
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows (as this isn't supported by PyTorch). Note also
that since this command runs without privillege the "system" method is
available only on Windows.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow \code{reticulate} to
automatically find an appropriate \code{conda} binary. See \strong{Finding Conda} for more details.}

\item{version}{PyTorch version to install. Specify "default" to install
  the CPU version of the latest release. Specify "gpu" to install the GPU
  version of the latest release.

  You can also provide a full major.minor.patch specification (e.g. "1.1.0"),
  appending "-gpu" if you want the GPU version (e.g. "1.1.0-gpu").

  Alternatively, you can provide the full URL to an installer binary (e.g.
  for a nightly binary).}

\item{envname}{Name of Python environment to install within}

\item{extra_packages}{Additional Python packages to install along with
PyTorch.}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}

\item{conda_python_version}{the python version installed in the created conda
environment. Python 3.6 is installed by default.}

\item{pip}{logical}

\item{channel}{conda channel}

\item{cuda_version}{string for the cuda toolkit version}

\item{dry_run}{logical, set to TRUE for unit tests, otherwise will run commands}

\item{...}{other arguments passed to [reticulate::conda_install()] or
[reticulate::virtualenv_install()].}
}
\description{
Install PyTorch and its dependencies
}
