% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_torch_extras}
\alias{install_torch_extras}
\title{Install additional Python packages alongside TensorFlow}
\usage{
install_torch_extras(packages, conda = "auto")
}
\arguments{
\item{packages}{Python packages to install}

\item{conda}{Path to conda executable (or "auto" to find conda using the PATH
and other conventional install locations). Only used when TensorFlow is
installed within a conda environment.}
}
\description{
This function is deprecated. Use the `extra_packages` argument to
`install_tensorflow()` to install additional packages.
}
