% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{length.torch.Tensor}
\alias{length.torch.Tensor}
\title{Length of a tensor.}
\usage{
\method{length}{torch.Tensor}(x)
}
\arguments{
\item{x}{tensor}
}
\value{
the number of elements of a tensor as an integer
}
\description{
This function is equivalent to torch$numel()
}
\examples{
\dontrun{
uo = torch$ones(3L, 5L)   # tensor with 15 elements
length(uo)
}

}
