% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{log.torch.Tensor}
\alias{log.torch.Tensor}
\title{Logarithm of a tensor given the tensor and the base}
\usage{
\method{log}{torch.Tensor}(x, base = exp(1L))
}
\arguments{
\item{x}{a tensor}

\item{base}{the base of the logarithm}
}
\description{
Logarithm of a tensor given the tensor and the base
}
\examples{
\donttest{
x <- torch$tensor(c(512, 1024, 2048, 4096))   # tensor([ 9., 10., 11., 12.])
base <- 2
log(x, base)

x <- torch$tensor(c(1, 10, 100, 1000))         # tensor([0., 1., 2., 3.])
log(x, 10)
}
}
