% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{log2.torch.Tensor}
\alias{log2.torch.Tensor}
\title{Logarithm of a tensor in base 2}
\usage{
\method{log2}{torch.Tensor}(x)
}
\arguments{
\item{x}{a tensor}
}
\description{
Logarithm of a tensor in base 2
}
