% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{logical_not}
\alias{logical_not}
\alias{!.torch.Tensor}
\title{Logical NOT of a tensor}
\usage{
\method{!}{torch.Tensor}(x)
}
\arguments{
\item{x}{tensor}
}
\value{
A tensor of booleans, where False corresponds to 0, and 1 to True
in a tensor of data type \code{torch$bool}.
}
\description{
There is not equivalent function in PyTorch for this generic.
To generate This generic we use the function \code{np$logical_not(x)}.
}
\examples{
\donttest{
A <- torch$ones(5L)
!A

Z <- torch$zeros(5L)
!Z
}
}
