% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{make_copy}
\alias{make_copy}
\title{Make copy of tensor, numpy array or R array}
\usage{
make_copy(object, ...)
}
\arguments{
\item{object}{a torch tensor or numpy array or R array}

\item{...}{additional parameters}
}
\description{
A copy of an array or tensor might be needed to prevent warnings
by new PyTorch versions on overwriting the numpy object
}
