% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{torch}
\alias{torch}
\alias{np}
\alias{torchvision}
\title{Main PyTorch module}
\format{PyTorch module}
\usage{
torch

np

torchvision
}
\description{
Interface to main PyTorch module. Provides access to top level classes

Interface to numpy module.

Interface to Torchvision module.
}
\keyword{datasets}
