% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{import}
\alias{import_builtins}
\alias{import_from_path}
\alias{dict}
\alias{tuple}
\alias{np_array}
\alias{array_reshape}
\alias{iterate}
\alias{py_len}
\alias{\%as\%}
\alias{iter_next}
\alias{py_run_string}
\alias{py_eval}
\alias{r_to_py}
\alias{py_to_r}
\alias{py_get_item}
\alias{use_python}
\alias{use_virtualenv}
\alias{use_condaenv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate:with-as-operator]{\%as\%}}, \code{\link[reticulate]{array_reshape}}, \code{\link[reticulate]{dict}}, \code{\link[reticulate]{import}}, \code{\link[reticulate:import]{import_builtins}}, \code{\link[reticulate:import]{import_from_path}}, \code{\link[reticulate:iterate]{iter_next}}, \code{\link[reticulate]{iterate}}, \code{\link[reticulate]{np_array}}, \code{\link[reticulate:py_run]{py_eval}}, \code{\link[reticulate]{py_get_item}}, \code{\link[reticulate]{py_len}}, \code{\link[reticulate:py_run]{py_run_string}}, \code{\link[reticulate:r-py-conversion]{py_to_r}}, \code{\link[reticulate:r-py-conversion]{r_to_py}}, \code{\link[reticulate]{tuple}}, \code{\link[reticulate:use_python]{use_condaenv}}, \code{\link[reticulate]{use_python}}, \code{\link[reticulate:use_python]{use_virtualenv}}}
}}

