% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTorch-package.R
\docType{package}
\name{rTorch-package}
\alias{rTorch}
\alias{rTorch-package}
\title{rTorch: R bindings to PyTorch}
\description{
R implementation of Machine Learning platform developed in Python.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f0nzie/rTorch}
}

}
\author{
\strong{Maintainer}: Alfonso R. Reyes \email{alfonso.reyes@oilgainsanalytics.com}

}
\keyword{internal}
