% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{tensor_ops}
\alias{tensor_ops}
\alias{+.torch.Tensor}
\alias{-.torch.Tensor}
\alias{*.torch.Tensor}
\alias{/.torch.Tensor}
\alias{==.torch.Tensor}
\alias{!=.torch.Tensor}
\alias{<.torch.Tensor}
\alias{<=.torch.Tensor}
\alias{>.torch.Tensor}
\alias{>=.torch.Tensor}
\alias{^.torch.Tensor}
\alias{&.torch.Tensor}
\alias{|.torch.Tensor}
\title{Two tensor operations}
\usage{
tensor_ops(a, b)

\method{+}{torch.Tensor}(a, b)

\method{-}{torch.Tensor}(a, b)

\method{*}{torch.Tensor}(a, b)

\method{/}{torch.Tensor}(a, b)

\method{==}{torch.Tensor}(a, b)

\method{!=}{torch.Tensor}(a, b)

\method{<}{torch.Tensor}(a, b)

\method{<=}{torch.Tensor}(a, b)

\method{>}{torch.Tensor}(a, b)

\method{>=}{torch.Tensor}(a, b)

\method{^}{torch.Tensor}(a, b)

\method{&}{torch.Tensor}(a, b)

\method{|}{torch.Tensor}(a, b)
}
\arguments{
\item{a}{tensor}

\item{b}{tensor}
}
\description{
Two tensor operations
}
\section{Methods (by class)}{
\itemize{
\item \code{torch.Tensor}: Add two tensors

\item \code{torch.Tensor}: Subtract two tensors

\item \code{torch.Tensor}: Tensor multiplication

\item \code{torch.Tensor}: Divide two tensors

\item \code{torch.Tensor}: Compare two tensors if equal

\item \code{torch.Tensor}: Compare two tensors if not equal

\item \code{torch.Tensor}: A tensor less than another tensor

\item \code{torch.Tensor}: A tensor less or equal than another tensor

\item \code{torch.Tensor}: A tensor greater than another tensor

\item \code{torch.Tensor}: A tensor greater or equal than another tensor

\item \code{torch.Tensor}: A tensor 'a' to the power of 'b'

\item \code{torch.Tensor}: Logical AND of two tensors

\item \code{torch.Tensor}: Logical OR of two tensors
}}

\examples{
\dontrun{
a <- torch$Tensor(list(1, 1, 1))
b <- torch$Tensor(list(2, 2, 2))
s <- 2.0
a + b
b - a
a * b
a / s
a == b
a == a
a != a
x <- torch$Tensor(list(list(2, 2, 2), list(4, 4, 4)))
y <- torch$Tensor(list(list(1, 2, 1), list(3, 4, 5)))
x > y
x < y
x >= y
y <= x
diag <- torch$eye(3L)
zeros <- torch$zeros(c(3L, 3L))
diag & zeros
diag & diag
diag | diag
zeros | zeros
zeros & zeros
diag & zeros
diag | zeros
}
}
