% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{torch_size}
\alias{torch_size}
\title{Size of a torch tensor object}
\usage{
torch_size(obj)
}
\arguments{
\item{obj}{a torch tensor object}
}
\description{
Get the size of a torch tensor or of torch.size object
}
