% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{all_dims}
\alias{all_dims}
\title{All dims}
\usage{
all_dims()
}
\description{
This function returns an object that can be used when subsetting tensors with
\code{[}. If you are familiar with Python, this is equivalent to the Python Ellipsis
\code{...}, (not to be confused with \code{...} in \code{R}). In Python, if \code{x} is a numpy
array or a torch tensor, in \code{x[..., i]} the ellipsis means "expand to match number of
dimension of x".
To translate the above Python expression to R, write:
\code{x[all_dims(), i]}.
}
\examples{
\donttest{
# Run this
d <- torch$tensor(list(list(0, 0),
                       list(0, 0),
                       list(0, 1),
                       list(1, 1)), dtype=torch$uint8)
d[all_dims(), 1]

f <- torch$arange(9L)$reshape(c(3L, 3L))
f
f[all_dims()]
f[all_dims(), 1L]
}
}
