% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{conda_install}
\alias{conda_install}
\alias{conda_python}
\title{Conda install function (modified from reticulate)}
\usage{
conda_install(
  envname = NULL,
  packages,
  forge = FALSE,
  pip = FALSE,
  pip_ignore_installed = TRUE,
  conda = "auto",
  python_version = NULL,
  channel = NULL,
  ...
)

conda_python(envname = NULL, conda = "auto")
}
\arguments{
\item{envname}{Name of conda environment}

\item{packages}{Character vector with package names to install or remove.}

\item{forge}{Include the  \href{https://conda-forge.org/}{Conda Forge} repository.}

\item{pip}{\code{TRUE} to use pip (defaults to \code{FALSE})}

\item{pip_ignore_installed}{Ignore installed versions when using pip. This is \code{TRUE} by default
so that specific package versions can be installed even if they are downgrades. The \code{FALSE}
option is useful for situations where you don't want a pip install to attempt an overwrite
of a conda binary package (e.g. \emph{SciPy} on \emph{Windows} which is very difficult to install via
\code{pip} due to compilation requirements).}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{python_version}{is the \emph{Python} version to install. Python \strong{3.4} is
installed by default. But you could specify for instance:
\code{python_version="3.7"}.}

\item{channel}{conda channel for pytorch}

\item{...}{Optional arguments, reserved for future expansion.}
}
\value{
\code{conda_list()} returns a data frame with the names and paths to the
respective python binaries of available environments. \code{conda_create()}
returns the Path to the python binary of the created environment.
\code{conda_binary()} returns the location of the main conda binary or \code{NULL}
if none can be found.
}
\description{
R functions for managing Python \href{https://conda.io/docs/user-guide/tasks/manage-environments.html}{conda environments}.
}
