% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dim.torch.Tensor}
\alias{dim.torch.Tensor}
\title{Dimensions of a tensor}
\usage{
\method{dim}{torch.Tensor}(x)
}
\arguments{
\item{x}{tensor}
}
\value{
a vector of integers with the dimensions of the tensor
}
\description{
Get the dimensions of a tensor displaying it as a vector.
}
