% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{\%.*\%}
\alias{\%.*\%}
\title{Dot product of two tensors
PyTorch dot function}
\usage{
a \%.*\% b
}
\arguments{
\item{a}{Tensor 1}

\item{b}{Tensor 2}
}
\description{
Dot product of two tensors
PyTorch dot function
}
