% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{\%**\%}
\alias{\%**\%}
\title{Matrix/Tensor multiplication of two tensors
PyTorch matmul}
\usage{
a \%**\% b
}
\arguments{
\item{a}{Tensor 1}

\item{b}{Tensor 2}
}
\description{
Matrix/Tensor multiplication of two tensors
PyTorch matmul
}
