% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{log.torch.Tensor}
\alias{log.torch.Tensor}
\title{Logarithm of a tensor given the tensor and the base}
\usage{
\method{log}{torch.Tensor}(x, base = exp(1L))
}
\arguments{
\item{x}{a tensor}

\item{base}{the base of the logarithm}
}
\description{
Logarithm of a tensor given the tensor and the base
}
