% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{log10.torch.Tensor}
\alias{log10.torch.Tensor}
\title{Logarithm of a tensor in base 10}
\usage{
\method{log10}{torch.Tensor}(x)
}
\arguments{
\item{x}{a tensor}
}
\description{
Logarithm of a tensor in base 10
}
\examples{
\donttest{
x <- torch$tensor(c(1, 10, 100, 1000))     # tensor([0., 1., 2., 3.])
}
}
