% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{one_tensor_op}
\alias{one_tensor_op}
\alias{exp.torch.Tensor}
\title{One tensor operation}
\usage{
one_tensor_op(x)

\method{exp}{torch.Tensor}(x)
}
\arguments{
\item{x}{tensor}
}
\description{
One tensor operation
}
\section{Methods (by class)}{
\itemize{
\item \code{torch.Tensor}: Exponential of a tensor
}}

\examples{
\dontrun{
A <- torch$ones(c(60000L, 1L, 28L, 28L))
dim(A)
}
}
