% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{import}
\alias{reexports}
\alias{dict}
\alias{reexports}
\alias{tuple}
\alias{reexports}
\alias{np_array}
\alias{reexports}
\alias{array_reshape}
\alias{reexports}
\alias{iterate}
\alias{reexports}
\alias{iter_next}
\alias{reexports}
\alias{import_builtins}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate]{import}}, \code{\link[reticulate]{dict}}, \code{\link[reticulate]{tuple}}, \code{\link[reticulate]{np_array}}, \code{\link[reticulate]{array_reshape}}, \code{\link[reticulate]{iterate}}, \code{\link[reticulate]{iter_next}}, \code{\link[reticulate]{import_builtins}}}
}}

