% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{torch_config}
\alias{torch_config}
\alias{torch_version}
\title{Torch configuration information}
\usage{
torch_config()

torch_version()
}
\value{
List with information on the current configuration of PyTorch.
You can determine whether PyTorch was found using the \code{available}
member (other members vary depending on whether \code{available} is \code{TRUE}
or \code{FALSE})
}
\description{
Torch configuration information
}
\keyword{internal}
